Face Detection Unit
Built a face detection system using OpenCV.
Used rapid object detection which uses the boosted cascade of the classifier by Viola-Jones that works with the Haar-like features. The object detection method returns the x- and y- coordinates, height, and width of the rectangle boxed in the face image.
As only the eyes and mouth part of the face are needed for the feature set. So the image was scaled down to 100x100 pixels and then a 80x30 rectangular pixels window was used for detecting eyes and a 50x40 pixel window was used for detecting mouth.